+++ talktive_client/lib/src/protocol/resident.dart
@@ -47,8 +47,12 @@ abstract class Resident implements _i1.SerializableModel {
     bool? showTypingIndicator,
     bool? showVoiceMessages,
     bool? showNeighborsDiscovery,
+    bool? allowDiscovery,
     bool? showCustomAvatar,
     bool? showImagesInPlaza,
+    bool? showImagesInLounges,
+    bool? showImagesInPrivateChats,
+    bool? showImagesInMoments,
     bool? showOthersOnlineStatus,
     bool? showOthersReadReceipts,
     bool? showOthersTypingIndicators,
@@ -69,8 +73,12 @@ abstract class Resident implements _i1.SerializableModel {
        showTypingIndicator = showTypingIndicator ?? true,
        showVoiceMessages = showVoiceMessages ?? true,
        showNeighborsDiscovery = showNeighborsDiscovery ?? true,
+       allowDiscovery = allowDiscovery ?? true,
        showCustomAvatar = showCustomAvatar ?? true,
        showImagesInPlaza = showImagesInPlaza ?? true,
+       showImagesInLounges = showImagesInLounges ?? true,
+       showImagesInPrivateChats = showImagesInPrivateChats ?? true,
+       showImagesInMoments = showImagesInMoments ?? true,
        showOthersOnlineStatus = showOthersOnlineStatus ?? true,
        showOthersReadReceipts = showOthersReadReceipts ?? true,
        showOthersTypingIndicators = showOthersTypingIndicators ?? true,
@@ -108,8 +116,12 @@ abstract class Resident implements _i1.SerializableModel {
     bool? showTypingIndicator,
     bool? showVoiceMessages,
     bool? showNeighborsDiscovery,
+    bool? allowDiscovery,
     bool? showCustomAvatar,
     bool? showImagesInPlaza,
+    bool? showImagesInLounges,
+    bool? showImagesInPrivateChats,
+    bool? showImagesInMoments,
     bool? showOthersOnlineStatus,
     bool? showOthersReadReceipts,
     bool? showOthersTypingIndicators,
@@ -207,6 +219,9 @@ abstract class Resident implements _i1.SerializableModel {
           : _i1.BoolJsonExtension.fromJson(
               jsonSerialization['showNeighborsDiscovery'],
             ),
+      allowDiscovery: jsonSerialization['allowDiscovery'] == null
+          ? null
+          : _i1.BoolJsonExtension.fromJson(jsonSerialization['allowDiscovery']),
       showCustomAvatar: jsonSerialization['showCustomAvatar'] == null
           ? null
           : _i1.BoolJsonExtension.fromJson(
@@ -216,6 +231,22 @@ abstract class Resident implements _i1.SerializableModel {
           ? null
           : _i1.BoolJsonExtension.fromJson(
               jsonSerialization['showImagesInPlaza'],
+            ),
+      showImagesInLounges: jsonSerialization['showImagesInLounges'] == null
+          ? null
+          : _i1.BoolJsonExtension.fromJson(
+              jsonSerialization['showImagesInLounges'],
+            ),
+      showImagesInPrivateChats:
+          jsonSerialization['showImagesInPrivateChats'] == null
+          ? null
+          : _i1.BoolJsonExtension.fromJson(
+              jsonSerialization['showImagesInPrivateChats'],
+            ),
+      showImagesInMoments: jsonSerialization['showImagesInMoments'] == null
+          ? null
+          : _i1.BoolJsonExtension.fromJson(
+              jsonSerialization['showImagesInMoments'],
             ),
       showOthersOnlineStatus:
           jsonSerialization['showOthersOnlineStatus'] == null
@@ -309,9 +340,17 @@ abstract class Resident implements _i1.SerializableModel {
 
   bool showNeighborsDiscovery;
 
+  bool allowDiscovery;
+
   bool showCustomAvatar;
 
   bool showImagesInPlaza;
+
+  bool showImagesInLounges;
+
+  bool showImagesInPrivateChats;
+
+  bool showImagesInMoments;
 
   bool showOthersOnlineStatus;
 
@@ -358,8 +397,12 @@ abstract class Resident implements _i1.SerializableModel {
     bool? showTypingIndicator,
     bool? showVoiceMessages,
     bool? showNeighborsDiscovery,
+    bool? allowDiscovery,
     bool? showCustomAvatar,
     bool? showImagesInPlaza,
+    bool? showImagesInLounges,
+    bool? showImagesInPrivateChats,
+    bool? showImagesInMoments,
     bool? showOthersOnlineStatus,
     bool? showOthersReadReceipts,
     bool? showOthersTypingIndicators,
@@ -402,8 +445,12 @@ abstract class Resident implements _i1.SerializableModel {
       'showTypingIndicator': showTypingIndicator,
       'showVoiceMessages': showVoiceMessages,
       'showNeighborsDiscovery': showNeighborsDiscovery,
+      'allowDiscovery': allowDiscovery,
       'showCustomAvatar': showCustomAvatar,
       'showImagesInPlaza': showImagesInPlaza,
+      'showImagesInLounges': showImagesInLounges,
+      'showImagesInPrivateChats': showImagesInPrivateChats,
+      'showImagesInMoments': showImagesInMoments,
       'showOthersOnlineStatus': showOthersOnlineStatus,
       'showOthersReadReceipts': showOthersReadReceipts,
       'showOthersTypingIndicators': showOthersTypingIndicators,
@@ -453,8 +500,12 @@ class _ResidentImpl extends Resident {
     bool? showTypingIndicator,
     bool? showVoiceMessages,
     bool? showNeighborsDiscovery,
+    bool? allowDiscovery,
     bool? showCustomAvatar,
     bool? showImagesInPlaza,
+    bool? showImagesInLounges,
+    bool? showImagesInPrivateChats,
+    bool? showImagesInMoments,
     bool? showOthersOnlineStatus,
     bool? showOthersReadReceipts,
     bool? showOthersTypingIndicators,
@@ -492,8 +543,12 @@ class _ResidentImpl extends Resident {
          showTypingIndicator: showTypingIndicator,
          showVoiceMessages: showVoiceMessages,
          showNeighborsDiscovery: showNeighborsDiscovery,
+         allowDiscovery: allowDiscovery,
          showCustomAvatar: showCustomAvatar,
          showImagesInPlaza: showImagesInPlaza,
+         showImagesInLounges: showImagesInLounges,
+         showImagesInPrivateChats: showImagesInPrivateChats,
+         showImagesInMoments: showImagesInMoments,
          showOthersOnlineStatus: showOthersOnlineStatus,
          showOthersReadReceipts: showOthersReadReceipts,
          showOthersTypingIndicators: showOthersTypingIndicators,
@@ -537,8 +592,12 @@ class _ResidentImpl extends Resident {
     bool? showTypingIndicator,
     bool? showVoiceMessages,
     bool? showNeighborsDiscovery,
+    bool? allowDiscovery,
     bool? showCustomAvatar,
     bool? showImagesInPlaza,
+    bool? showImagesInLounges,
+    bool? showImagesInPrivateChats,
+    bool? showImagesInMoments,
     bool? showOthersOnlineStatus,
     bool? showOthersReadReceipts,
     bool? showOthersTypingIndicators,
@@ -589,8 +648,13 @@ class _ResidentImpl extends Resident {
       showVoiceMessages: showVoiceMessages ?? this.showVoiceMessages,
       showNeighborsDiscovery:
           showNeighborsDiscovery ?? this.showNeighborsDiscovery,
+      allowDiscovery: allowDiscovery ?? this.allowDiscovery,
       showCustomAvatar: showCustomAvatar ?? this.showCustomAvatar,
       showImagesInPlaza: showImagesInPlaza ?? this.showImagesInPlaza,
+      showImagesInLounges: showImagesInLounges ?? this.showImagesInLounges,
+      showImagesInPrivateChats:
+          showImagesInPrivateChats ?? this.showImagesInPrivateChats,
+      showImagesInMoments: showImagesInMoments ?? this.showImagesInMoments,
       showOthersOnlineStatus:
           showOthersOnlineStatus ?? this.showOthersOnlineStatus,
       showOthersReadReceipts: